separate mex functions from C++ libraries
