+++ systems/robotInterfaces/@QPLocomotionPlanCPPWrapper/CMakeLists.txt
@@ -1,21 +1,20 @@
 include_directories(..)
 
-add_mex(getQPControllerInput getQPControllerInput.cpp)
-target_link_libraries(getQPControllerInput drakeQPLocomotionPlan drakeUtil)
+if (MATLAB_FOUND)
+  if (LCM_FOUND)
+    macro(add_qplocomotionplan_mex)
+      add_mex(${ARGV} ${ARGV}.cpp)
+      target_link_libraries(${ARGV} drakeQPLocomotionPlan drakeMexUtil)
+    endmacro()
 
-add_mex(start_time start_time.cpp)
-target_link_libraries(start_time drakeQPLocomotionPlan drakeUtil)
+    add_qplocomotionplan_mex(getQPControllerInput) 
+    add_qplocomotionplan_mex(start_time)
+    add_qplocomotionplan_mex(setStartTime)
+    add_qplocomotionplan_mex(getLastQPInput)
+    add_qplocomotionplan_mex(isFinished)
+    add_qplocomotionplan_mex(duration)
 
-add_mex(setStartTime setStartTime.cpp)
-target_link_libraries(setStartTime drakeQPLocomotionPlan drakeUtil)
+    add_subdirectory(private)
+  endif()
+endif()
 
-add_mex(getLastQPInput getLastQPInput.cpp)
-target_link_libraries(getLastQPInput drakeQPLocomotionPlan drakeUtil)
-
-add_mex(isFinished isFinished.cpp)
-target_link_libraries(isFinished drakeQPLocomotionPlan drakeUtil)
-
-add_mex(duration duration.cpp)
-target_link_libraries(duration drakeQPLocomotionPlan drakeUtil)
-
-add_subdirectory(private)
