+++ systems/robotInterfaces/@QPLocomotionPlanCPPWrapper/duration.cpp
@@ -1,5 +1,5 @@
 #include "QPLocomotionPlan.h"
-#include "drakeUtil.h"
+#include "drakeMexUtil.h"
 #include "drakeFloatingPointUtil.h"
 
 using namespace std;
