+++ systems/robotInterfaces/@QPLocomotionPlanCPPWrapper/private/CMakeLists.txt
@@ -1,4 +1,8 @@
 include_directories(../..)
 
-add_mex(setDuration setDuration.cpp)
-target_link_libraries(setDuration drakeQPLocomotionPlan drakeUtil)
+if (MATLAB_FOUND)
+  if (LCM_FOUND)
+    add_mex(setDuration setDuration.cpp)
+    target_link_libraries(setDuration drakeQPLocomotionPlan drakeMexUtil)
+  endif()
+endif()
