+++ systems/robotInterfaces/@QPLocomotionPlanCPPWrapper/private/setDuration.cpp
@@ -1,4 +1,4 @@
-#include "drakeUtil.h"
+#include "drakeMexUtil.h"
 #include "QPLocomotionPlan.h"
 
 using namespace std;
