+++ systems/robotInterfaces/@QPLocomotionPlanCPPWrapper/start_time.cpp
@@ -1,5 +1,5 @@
 #include "QPLocomotionPlan.h"
-#include "drakeUtil.h"
+#include "drakeMexUtil.h"
 #include "drakeFloatingPointUtil.h"
 
 using namespace std;
