+++ systems/robotInterfaces/CMakeLists.txt
@@ -30,14 +30,14 @@ if (MATLAB_FOUND)
     include_directories (${CMAKE_SOURCE_DIR}/util )
 
     add_mex(octomapWrapper octomapWrapper.cpp)
-    target_link_libraries(octomapWrapper drakeUtil)  # note: drakeUtil depends on eigen
+    target_link_libraries(octomapWrapper drakeUtil drakeMexUtil)  # note: drakeUtil depends on eigen
     pods_use_pkg_config_packages(octomapWrapper octomap)
 
   endif()
 
   if (LCM_FOUND)
     add_mex(constructQPLocomotionPlanmex constructQPLocomotionPlanmex.cpp)
-    target_link_libraries(constructQPLocomotionPlanmex drakeQPLocomotionPlan drakeUtil drakeZMPUtil)
+    target_link_libraries(constructQPLocomotionPlanmex drakeQPLocomotionPlan drakeUtil drakeZMPUtil drakeMexUtil)
 
     add_subdirectory(@QPLocomotionPlanCPPWrapper)
   endif()
